requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
soundfile==0.12.1
openai>=1.52.0
httpx>=0.25.0
